backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
openai>=1.0.0
upstash-redis>=0.3.0
python-dotenv>=1.0.0
nanoid>=2.0.0
pydantic>=2.0.0
truststore
